+++ src/test/java/com/sonarorw/rules/s1117/S1117_compliant_official.java
@@ -1,0 +1,9 @@
+package com.sonarorw.rules.s1117;
+
+class S1117_compliant_official {
+    private int value;
+
+    void setValue(int value) {
+        this.value = value;
+    }
+}