+++ src/test/java/com/sonarorw/rules/s1858/S1858_compliant_official.java
@@ -1,0 +1,7 @@
+package com.sonarorw.rules.s1858;
+
+class S1858_compliant_official {
+    boolean isFrance(String country) {
+        return "FR".equals(country);
+    }
+}