+++ src/test/java/com/sonarorw/rules/s2737/S2737_compliant_official.java
@@ -1,0 +1,11 @@
+package com.sonarorw.rules.s2737;
+
+class S2737_compliant_official {
+    void copy() throws Exception {
+        doSomething();
+    }
+
+    void doSomething() throws Exception {
+        throw new Exception("boom");
+    }
+}