+++ src/test/java/com/sonarorw/rules/s2959/S2959_compliant_official.java
@@ -1,0 +1,9 @@
+package com.sonarorw.rules.s2959;
+
+class S2959_compliant_official {
+    void printAll() {
+        for (int i = 0; i < 3; i++) {
+            System.out.println("Item" + i);
+        }
+    }
+}